pyproject.toml
[tool.poetry]
name = "flask-app"
version = "0.1.0"
description = ""
authors = ["António Alvarenga <aalvarenga@imaginarycloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.2"
gunicorn = "^20.1.0"
Flask-RESTful = "^0.3.9"
flask-apispec = "^0.11.4"
flask-sqlalchemy = "^3.0.2"
psycopg2-binary = "^2.9.4"
Flask-Migrate = "^3.1.0"
marshmallow-sqlalchemy = "^0.28.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
